While working on https://github.com/dotnet/runtime/issues/2191 I got surprised by how long it takes to get the benchmark results.

It turned out that so far we were always running the benchmarks one more time if `Inlining` or `TailCall` diagnosers were enabled. We simply don't need to do that, the overhead of these two diagnosers is very, very small (we just sign up for 3 events).

```cmd
git clone https://github.com/dotnet/BenchmarkDotNet/
cd BenchmarkDotNet
cd samples
cd BenchmarkDotNet.Samples.FSharp
dotnet run -c Release -f net461 --filter TailCallDetector
```

Before:

| Method | facRank |      Mean |     Error |    StdDev |
|------- |-------- |----------:|----------:|----------:|
|   test |       7 |  9.460 ns | 0.2158 ns | 0.2310 ns |
|  test1 |       7 |  4.758 ns | 0.0824 ns | 0.0731 ns |
|  test2 |       7 | 51.820 ns | 0.9814 ns | 0.9180 ns |

```log
Run time: 00:02:29 (149.11 sec), executed benchmarks: 3

Global total time: 00:02:31 (151.73 sec), executed benchmarks: 3
```

After:

| Method | facRank |      Mean |     Error |    StdDev |
|------- |-------- |----------:|----------:|----------:|
|   test |       7 |  9.383 ns | 0.0837 ns | 0.0783 ns |
|  test1 |       7 |  4.688 ns | 0.0655 ns | 0.0613 ns |
|  test2 |       7 | 49.748 ns | 0.2633 ns | 0.2199 ns |

```log
Run time: 00:01:23 (83.31 sec), executed benchmarks: 3

Global total time: 00:01:26 (86.03 sec), executed benchmarks: 3
```