@adamsitnik 2020-02-19
@GeorgePlotnikov thanks for a very fast response!

@AndreyAkinshin I need your approval to be able to merge